+++ application/src/main/java/common/param/ed/tr/TREvent.java
@@ -8,13 +8,13 @@ public class TREvent extends TxnEvent {
 
     private final String tweetID;
     private final String[] words;
-    private final int myBid;
+    private final double myBid;
     private final int myPid;
     private final String my_bid_array;
     private final String my_partition_index;
     private final int my_number_of_partitions;
 
-    public TREvent(int bid, int pid, String bid_array, String partition_index, int number_of_partitions,
+    public TREvent(double bid, int pid, String bid_array, String partition_index, int number_of_partitions,
                    String tweetID, String[] words) {
         super(bid, pid, bid_array, partition_index, number_of_partitions);
         this.myBid = bid;
@@ -26,7 +26,7 @@ public class TREvent extends TxnEvent {
         this.words = words;
     }
 
-    public int getMyBid() {
+    public double getMyBid() {
         return myBid;
     }
     public int getMyPid() {
@@ -51,7 +51,7 @@ public class TREvent extends TxnEvent {
     }
 
     public TREvent cloneEvent() {
-        return new TREvent((int) bid, pid, Arrays.toString(bid_array), Arrays.toString(partition_indexs), number_of_partitions, tweetID, words);
+        return new TREvent(bid, pid, Arrays.toString(bid_array), Arrays.toString(partition_indexs), number_of_partitions, tweetID, words);
     }
 
     @Override
